+++ SurveillanceTempsReel/Actors/HospitalEventFetcherActor.cs
@@ -26,7 +26,7 @@ namespace SurveillanceTempsReel.Actors
         private HashSet<IActorRef> _subscriptions;
         private ICancelable _cancelFetching;
         
-        private int _lastEventId;
+        private long _lastEventId;
 
         private readonly ILoggingAdapter _log = Logging.GetLogger( Context );
